requirements.txt
altgraph==0.17.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
idna==3.10
mysql-connector-python==9.3.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pygame==2.6.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMySQL==1.1.1
pyodbc==5.2.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
qrcode==8.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.32.4
shiboken6==6.9.1
urllib3==2.5.0
yt-dlp==2025.6.30
